requirements.txt
pymupdf
requests
numpy
pandas
pillow
tqdm
easyocr
python-dotenv
pydrive2
openpyxl
